requirements.txt
python_requires='>=3.7'
obs-websocket-py==5.0.1
pywin32>=223
pyinstaller==6.11.0